Update as9004_zoe_s.py
Added history questions and final answer code. Made it so that if you type the wrong number or a word it doesn't work and tells you to try again.  I did the same thing with whether you want to start the quiz or not so that it only accepts yes or no.